Fix user/group creation - the config file must be a package source
sddm user/group creation was broken since the build that dropped
the sysusers config file as a package source. We were still
trying to generate the %pre scriptlet from SOURCE18 even though
that didn't exist any more.

Unfortunately, there still appears to be no way to actually use
an upstream packaged config file as the target of the
%sysusers_create_compat macro. I posted to devel@ about this
back in 2021, which triggered an interesting thread, but no
immediate practical solution:
https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/thread/TFDMAU7KLMSQTKPJELHSM6PFVXIZ56GK/
so the only workaround I know of is, unfortunately, to duplicate
the config file as a package source :(

So this commit takes the generated file from a broken 0.20.0
build and adds it back as a package source. With this change,
the %pre script is once again correct and the package provides
`user(sddm)` and `group(sddm)` as it should.